backend/requirements.txt
# =====================================================
# DEPENDENCIAS DEL SISTEMA
# =====================================================

# Framework web
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Base de datos
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Autenticación y seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.2

# Validación y serialización
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilidades
python-dateutil==2.8.2
openpyxl==3.1.2
#pandas==2.1.4

# Manejo de archivos
aiofiles==23.2.1
python-magic==0.4.27

# Logging y monitoreo
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.26.0

# Desarrollo
black==24.1.1
flake8==7.0.0
mypy==1.8.0
